added useState feature to input values from forms
